move DeDotProcessor into its own plugin
