+++ plugins/ingest-dedot/src/main/java/org/elasticsearch/ingest/dedot/IngestDeDotPlugin.java
@@ -1,0 +1,48 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.elasticsearch.ingest.dedot;
+
+import org.elasticsearch.node.NodeModule;
+import org.elasticsearch.plugins.Plugin;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Files;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
+
+public class IngestDeDotPlugin extends Plugin {
+
+    @Override
+    public String name() {
+        return "ingest-dedot";
+    }
+
+    @Override
+    public String description() {
+        return "Ingest processor that replaces dots in field names";
+    }
+
+    public void onModule(NodeModule nodeModule) throws IOException {
+        nodeModule.registerProcessor(DeDotProcessor.TYPE, (templateService) -> new DeDotProcessor.Factory());
+    }
+}
