+++ plugins/ingest-dedot/src/test/resources/rest-api-spec/test/ingest_dedot/10_basic.yaml
@@ -1,0 +1,5 @@
+"Ingest plugin installed":
+    - do:
+        cluster.stats: {}
+
+    - match:  { nodes.plugins.0.name: ingest-dedot  }
