+++ plugins/ingest-dedot/src/test/resources/rest-api-spec/test/ingest_dedot/20_dedot_processor.yaml
